feat/move-to-main: messing around with structure
