+++ server/MainThreadHandler.ts
@@ -2,13 +2,14 @@ import { createStore } from 'redux'
 import indexReducer from './reducers/indexReducer';
 import { UPDATE_SETTINGS } from './reducers/settingsActions'
 import { loadSettings } from './utils/SettingsStorage'
-import { MainApp } from './components/MainApp'
 
 export class MainThreadHandlers {
     webContents: any
     store: any
 
     constructor(webContents: any) {
+        console.log('Creating MainThreadHandlers')
+
         this. webContents = webContents
         global.storeRedux = createStore(
             indexReducer
@@ -27,6 +28,8 @@ export class MainThreadHandlers {
     }
 
     ipcHandler() {
+        console.log('CALLING IPC HANDLER')
+
         // initiate 
         this.webContents.on('get-settings', (
             (event: any, payload: any) => { 
@@ -36,6 +39,5 @@ export class MainThreadHandlers {
         )
         this.webContents.send('get-settings', 'Start up ipcHandlers')
         this.webContents.send('to-renderer', 'Start up ipcHandlers')
-        let app = new MainApp(this.webContents)
     }
 }
