+++ server/preload.js
@@ -13,7 +13,6 @@ global.emberplus = remote.require('emberplus')
 global.casparcgconnection = remote.require('casparcg-connection')
 global.dialog = dialog
 global.getPath = getPath
-//global.SettingsStorage = './utils/SettingsStorage.ts'
 
 
 console.log('PreLoad executed')
